backend/requirements.txt
mysql-connector-python==8.0.33
Flask==2.3.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
